refactor: 유틸리티 클래스 제거
