+++ src/main/java/chess/domain/gamestate/State.java
@@ -2,6 +2,11 @@ package chess.domain.gamestate;
 
 import chess.domain.board.Board;
 import chess.domain.dto.ScoreDto;
+import chess.domain.gamestate.finished.End;
+import chess.domain.gamestate.running.Move;
+import chess.domain.gamestate.running.Ready;
+import chess.domain.gamestate.running.Start;
+import chess.domain.gamestate.running.Status;
 import chess.domain.team.Team;
 
 public interface State {
@@ -21,4 +26,23 @@ public interface State {
     ScoreDto judgeResult();
 
     boolean isAnyKingDead();
+
+    static State generateState(String stateSignature, Board board) {
+        if (stateSignature.equals("ready")) {
+            return new Ready(board);
+        }
+        if (stateSignature.equals("start")) {
+            return new Start(board);
+        }
+        if (stateSignature.equals("move")) {
+            return new Move(board);
+        }
+        if (stateSignature.equals("end")) {
+            return new End(board);
+        }
+        if (stateSignature.equals("status")) {
+            return new Status(board);
+        }
+        throw new IllegalArgumentException("[ERROR] 존재하지 않는 명령입니다.");
+    }
 }
