+++ src/main/java/chess/domain/piece/Piece.java
@@ -28,6 +28,30 @@ public abstract class Piece {
         this.location = Location.of(location.getX(), location.getY());
     }
 
+    public static Piece generatePiece(long id, long roomId, char signature, String team, String location) {
+        Team pieceTeam = Team.of(team);
+        Location pieceLocation = Location.of(location);
+        if (signature == 'r') {
+            return Rook.of(id, roomId, pieceTeam, pieceLocation);
+        }
+        if (signature == 'k') {
+            return King.of(id, roomId, pieceTeam, pieceLocation);
+        }
+        if (signature == 'q') {
+            return Queen.of(id, roomId, pieceTeam, pieceLocation);
+        }
+        if (signature == 'p') {
+            return Pawn.of(id, roomId, pieceTeam, pieceLocation);
+        }
+        if (signature == 'n') {
+            return Knight.of(id, roomId, pieceTeam, pieceLocation);
+        }
+        if (signature == 'b') {
+            return Bishop.of(id, roomId, pieceTeam, pieceLocation);
+        }
+        throw new IllegalArgumentException("[ERROR] 존재하지 않는 기물입니다.");
+    }
+
     public void moveTo(Location target, Board board) {
         validateTargetIsNotSameTeam(target, board);
         validateMovingAbilityToTarget(target);
